PR feedback and linter again
